md_monitor: rescan array when rebuild started
We should rescan the array once we receive a 'RebuildStarted'
message, as this already indicates that something has changed
with the array layout.
At that point the new disk is already attached, so we will be
able to get the correct index and slot enumeration.
And we can track precisely when the new disk becomes active,
avoiding any race conditions upon 'SpareActive' messages.

References: bnc#787819

Signed-off-by: Hannes Reinecke <hare@suse.de>