md_monitor: do not set the slot number to -1
mdadm will set the slot number to '-1' if a device is being
removed as part of the recovery process.
However, that causes md_monitor to lose information about which
side this device originally resided in.
So md_monitor should not update the slot number to '-1'; if the
number changed it'll be updated once the new device has been
activated.

References: bnc#799939

Signed-off-by: Hannes Reinecke <hare@suse.de>