md_monitor: Set DASD timeout flag prior to calling mdadm
mdadm relies on all I/O to be aborted before any changes to
the array can be made. So whenever we detect a timeout
situation we should be setting the DASD timeout flag
before calling mdadm. This will ensure that all I/O will be
aborted in a timely manner.

References: bnc#808302

Signed-off-by: Hannes Reinecke <hare@suse.de>