md_monitor: Do not fail devices during discovery
When doing a discovery the aio_timeout will be set to '0' to
get an initial state update from MD. However, the I/O status
will be set to 'IO_PENDING' (as the checker hasn't run yet).
This does not constitute an error, so we shouldn't try to
fail the mirror here.

References: bnc#828121

Signed-off-by: Hannes Reinecke <hare@suse.de>