md_monitor: Do not fail devices on z/VM 'detach'
When a 'move' event is received the device itself will only
start reporting I/O errors a short while after. So if we
start to fail devices pre-emptively on this event the path
checker will detect the devices as running. This will cause
a recovery to start, which will immediately being interrupted
by the device throwing I/O errors.

So we should rather ignore this event and let md_monitor catch
the resulting I/O errors.

Signed-off-by: Hannes Reinecke <hare@suse.de>