monitor_testcase_6: Cleanups
When waiting for reservations to be registered we should be
limiting the wait, so as not to wait forever.
And the call to 'wget' should be put into a function for
better readability.
And we should always set the device to 'online' via wget
to be sure the reserve call will succeed.

Signed-off-by: Hannes Reinecke <hare@suse.de>