monitor_testcase_11: Do not print 'failed array' message
After having stopped the array it is expected that any call
to mdadm or md_monitor will return an error due to the array
not being available. So instead of printing out confusing
messages we should just check if the array is indeed stopped
and return an error if it isn't.

Signed-off-by: Hannes Reinecke <hare@suse.de>