monitor_testcase_2: Add 'wait_md' after failing devices
Calling 'mdadm --fail <list of devices>' has the unfortunate
effect that mdadm will be calling 'Fail' for each device in turn.
This will confuse md_monitor as it'll start recovery for the first
device, then get a 'Fail' request for the second, and start failing
the entire array _again_.

So we should rather wait for md to finish the 'failing' sequence
before proceeding further.

Signed-off-by: Hannes Reinecke <hare@suse.de>