monitor_testcase_3: use 'MonitorStatus' instead of 'ArrayStatus'
ArrayStatus prints out the full description of the array, within
which the order of the devices is basically arbitrary. In particular
it might change after reassembling, causing random failures.
As we're only interested that md_monitor picked up every device
we can just compare the 'MonitorStatus' CLI command which
doesn't suffer from this issue.

Signed-off-by: Hannes Reinecke <hare@suse.de>