monitor_testcase_11: Fix random failures during reassembly
md_monitor might be slow to pick up re-assembled devices,
as it waits for the udev event to arrive.
This might be delayed as udev has to scan the new device etc.
So we need to add a waiting loop when checking the md_monitor
status and only terminate the testcase if md_monitor doesn't
register the new array after a certain timeout.

Signed-off-by: Hannes Reinecke <hare@suse.de>