monitor_setup_root: Initial commit
This is a script to take the currently running system and copy it over to newly
created MD.  As part of the quest, separate disks for /boot and /var/log are
set up.  The disk id's for /boot and /var/log are currently hardcoded.  The
script also assumes the currently running kernel will be running from MD, which
is fine for the type of testing we perform.

Signed-off-by: Libor Pechacek <lpechacek@suse.cz>