monitor_testcase: work around bug#954769
Newer versions of mdadm might reorder the devices in the output
of 'mdadm --detail'. The testcases would do a simple 'diff' to
compare the output, leading to spurious failures.
So this patch introduces a common function 'check_md_log' which
provides for a workaround by simply only checking the lines we're
interested in.

References: bsc#954769

Signed-off-by: Hannes Reinecke <hare@suse.de>