md_monitor: Ignore inactive arrays
Due to udev event sequencing the 'change' event for the MD array
might arrive when the array is still stopped. This we deliver us
a partial uevent with most of the values missing, causing md_monitor
to not correctly detect the array.
Looking at the kernel source an array is inactive if no personality
is set and either the number of disks is 0 or the size is 0.
As we're already testing for the first two conditions we only need
to check for the size to figure out if the array is inactive.

References: bsc#1068175

Signed-off-by: Hannes Reinecke <hare@suse.com>