md_monitor: store alias in struct md_dev to avoid lookup errors
An MD device can have a device alias instead of the usual /dev/mdXXX
device node. As udev might have a lag when processing events we
might not be able to always get the alias from the uevent variables.
So store the alias in the 'dev_name' field to avoid lookup errors.

References: bsc#1068175

Signed-off-by: Hannes Reinecke <hare@suse.com>