testcases/monitor_testcase_functions: work around dirty bitmap
Upon exit the MD bitmap might still be dirty, but the fix from
bsc#951211 is not implemented in SLES11.
So use the workaround given in the bug, namely writing '0' to
the 'time_base' sysfs attribute and wait for two seconds to
give MD a chance to clear the bitmap.

References: bsc#951211

Signed-off-by: Hannes Reinecke <hare@suse.com>