md_monitor: new thread to repeat failed array reset
Sometimes "fail" and "reset" actions may happen in very quick
succession. The kernel thread that is responsible for recovery
actions (raid10d) needs some time to bring the array in a clean
status after a "mdadm --fail" command. If a "--re-add faulty"
command is sent before this state is reached, the command may
fail, causing devices not to be brought back online although
they are healthy. The "--re-add faulty" command must be retried
in this case.

This patch creates a new thread that handles this situation.
It is only started after a reset command fails, and does nothing
but wait for some time and then put the array's reset action
back on the pending queue.