fail_mirror: don't set io_status = IO_OK
The io_status is normally only set by the device monitor thread.
Blindly setting it to IO_OK here (in the code path where the
device has been marked as failed by MD) seems too optimistic,
even in the unlikely case that the md status is now IN_SYNC.

Rather, set it to IO_UNKNOWN and notify the device monitor
to re-check. This means that we may have to wait for the check
to complete in reset_mirror(). But if the device is really healthy,
we won't need to wait long, so limit the wait to a few ms.