md_monitor: permanently store 'md_side' for each device
A device might be removed or re-added, but in either case the side
at which this happens should not change. So we can treat the md_side
setting as persistent, and decouple it from the md_slot logic.
This is especially helpful as md_slot will change during recovery.

References: bsc#1116560

Signed-off-by: Hannes Reinecke <hare@suse.com>