md_monitor: save original slot position for display
The md_slot variable will be set to '-1' when a device been faulted,
leaving us with no information where this device was originally located
at. This results in the status display not able to print the correct
state for each device.
This patch introduces an additional variable 'md_slot_saved', holding
the original slot number. With this we have the slot position avaiable
and can print out the correct status.

Signed-off-by: Hannes Reinecke <hare@suse.com>