md_monitor: Do not set 'REMOVED' state if ioctl fails
When the ioctl in md_rdev_check_state() we should not set the
device to REMOVED, but rather to UNKNOWN. REMOVED should be
reserved for cases where the MD array indicates that this device
has been removed.
An ioctl failure just indicates that we cannot talk to the array,
hence UNKNOWN is the correct state here.

References: bsc#1125281

Signed-off-by: Hannes Reinecke <hare@suse.com>