md_monitor: display_md_status: don't omit trailing dots
display_md_status() and display_io_status() end
output at the index of the last child of the md array they encounter
while walking the sibling list. This may lead to missing trailing dots (".")
if a disk at the end of the array was in removed.

This behavior was introduced by the past commits mentioned below,
with the intention to fix a crash that occured when a degraded
md_monitor struct (with md_dev->raid_disks == -1) was encountered.
Fix this case by using md_dev->raid_disks again if it is larger
than the last slot encountered.

Fixes: 2379a9c md_monitor: fixup crash in display_io_status
Dixes: 62d385b md_monitor: fixup crash in display_md_status